Alarms Web Page brings down webadmin app if store.xml not found
If the default XML implementation is used, the web page which must access the alarms store waits to see the correct file defined by alarms.store.path.  If after a short interval it does not find it, it throws an exception.  This will bring down the entire Webadmin app, unfortunately.

It is perfectly legitimate for the file not to be there if the user has chosen not to run the alarm service, which produces it.  The behavior should be to exit gracefully from doing anything.

This patch allows for an initialized state.  If the file is not found, the error is reported but any further query functionality is disabled.

Tested on testbed deployment.  Note that a separate patch will be required for 2.5.

Target: trunk
Patch: http://rb.dcache.org/r/5258
Require-book: no
Require-notes: yes
Acked-by: Tigran


RELEASE NOTES: fixes a bug which crashes the webadmin app if the webapp page for alarms expects to find an xml store file but there is no alarm service producing it.