pool: Avoid CellInfo query for billing registration at end of transfer
CellInfo, in particular in pools, is relatively expensive to generate.
Thus using it at the end of every transfer should be avoided.

This patch refactors the request object to avoid the use of the low
level CellEndpoint interface. Instead the address of the pool and
communication stubs for billing and the door are injected through
the constructor.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>