pool: Simplify ChecksumMover interface
This patch is continuation of work started by Paul.

The patch simplifies the ChecksumMover interface implemented
by movers that are checksum aware. The interface is refined
to provide an expected checksum (as defined by the remote
end) and an actual checksum (as computed by the mover). The
pool can signal whether on-transfer checksum calculation is
enabled and specify a default checksum type to be used.

This is the first of a series of checksum related patches.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5310/