pnfsmanager: Remove PnfsSetStorageInfoMessage
Messages based on FileAttributes will slowly replace legacy
messages operations based on StorageInfo.

This patch removes PnfsSetStorageInfoMessage and replaces
the few places that used it with PnfsSetFileAttributes.

ChimeraNameSpaceProvider is extended to support setting
StorageInfo through FileAttributes. The access mode argument
is dropped as it was not used by Chimera.

The 'set storageinfo' command has been removed. The command
was limited to updating the access latency, and in a way that
looked incorrect. The command has been deprecated for a
while.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5321/