[fix] Add serialVersionUID field to legacy principals
After migrating to JGlobus 2, we added the old principals from
JGlobus 1 to ensure compatibility with older pools. For reasons
unknown to me, the serial version UID of the SimplePrincipal
class did not match the UID of the class shipped with JGlobus 1.
This prevent dCache 2.5/2.6 from working with dCache pool 2.2,
2.3 and 2.4.

This patch adds the version UID. Unfortunately this means that
dCache 2.5.2 will not be compatible with dCache 2.5.0 and
dCache 2.5.1. Given that we don't have any production installations
of dCache 2.5 yet, I consider compatibility with pre-2.5 releases
more important.

Target: trunk
Request: 2.5
Require-book: no
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>