pool: Simplify CellStubHelper
RepositorySubsystemTest uses CellStubHelper to mock an entire
cell system. CellStubHelper instantiates a SystemCell and an
actual cell for each cell that is mocked. This is fairly
expensive and complicates debugging when tests fail.

Since CellStubHelper was written, we have introduced the
CellEndpoint interface.

This patch more or less rewrites the CellStubHelper class to
avoid creating real cells. Instead CellEndpointHelper is a
minimal CellEndpoint implementation that redirects all
messages to CellStubHelper.

The changes to the test are deliberately kept minimal. I am
certain one could do cool things with Mockito, but this patch
is about making RepositorySubsystemTest easier to debug, not
to rewrite it completely.

Target: trunk
Require-book: no
Require-notes: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5331/