pool: Kill movers on shutdown
The SimpleIoScheduler has a shutdown method, but it doesn't actually
do anything.

This patch updates the shutdown method to kill all movers. Movers
are given a moment to die before the shutdown method returns. This
prevents that dependent components are shut down before movers
terminate.

An explicit dependency had to be added between the execution services
and the repository. This is to prevent that spring shuts down the
repository component before the movers are terminated.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5300/