poolmanager: Clarify use of preallocated space in pool selection
Pool manager can make use of the fie size or preallocated space
during write pool selection. The field is unused for all other
pool selections, yet the field is present in the base class. This
makes the purpose of the field less clear than it should be.

This patch moves the field to the write pool selection message
and the field is changed to precallocated to reflect its use. It
differs from the file size in storage info/file attributes by
being a preallocation, not an exact size of the final file. It is
only ever used when locating a pool with enough free space.

Some DCAP specific code is moved from pool manager to the DCAP
door.

The patch also makes file size and checksum attributes required
for read pool selection. These should be available to a pool
when staging a file or copying it between pools.

Target: trunk
Require-book: no
Require-notes: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5322