Deprecate access latency, retention policy and file size in StorageInfo
Some information can be represented in both StorageInfo and
FileAttributes. This duplication is unfortunate as it makes
the code much harder to read and is counter to the fine grained
granularity that was intended with the FileAttributes class.

This patch deprecates the access latency, retention policy and
file size fields in Storage Info. I cannot completely remove them
as:

- We need to maintain protocol compatibility with older pools

- The HSM scripts rely on those attributes as part of the
  storage info argument (which contains the toString() version
  of the StorageInfo).

- Pools have persistent copies of StorageInfo as part of their
  meta data.

I renamed the accessors to make it obvious that one should avoid
using them. They are now only used in some conversion code between
StorageInfo and FileAttributes, some unit tests, and a utility
to deserialize a StorageInfo from a file. For most of dCache,
these fields will not have been initialized.

Chimera is updated to decouple access latency and retention policy
from storage info.

The pool is updated to push FileAttributes all the way into the meta
data stores, although the stores will convert it to StorageInfo for
persistence. We need this to maintain backwards compatibility with
existing pools. The process is lossy, but all information we need is
maintained. I extended the flag-c map entry in StorageInfo to take
a comma-separated list of checksums.

Note that we documented some of the deprecated StorageInfo fields
as part of the billing file format strings. We did not use these
fields in our default patterns, but the change should be documented
in the release notes (ie removal of storageInfo.fileSize,
storageInfo.retentionPolicy, storageInfo.accessLatency). A later
patch will update the info messages to use file attributes, thus
reexposing the lost fields.

Dead code affected by the change is deleted.

Target: trunk
Require-book: no
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5324/