alarms: Add check for whether the alarms service is the only service in the domain
Since the alarms service loads an alternative logback configuraiton, it cannot
sensibly be mixed with other services. This patch adds a check in the batch
file which will refuse to start the alarms service if it is combined with other
services.

I had to make minor modifications to the boot loader (the old code was an
ugly hack anyway).

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: http://rb.dcache.org/r/5406/