pool: Get rid of PoolIORequest
A mover represents the pool side of a file transfer. It implements
the MoverProtocol interface. As part of scheduling and executing a
mover, the pool wraps the mover at least three times:

- PoolIOTransfer (and its subclasses) wrap the mover and the
  ReplicaDescriptor.

- PoolIORequest wraps PoolIOTransfer and adds billing and door
  notification.

- PrioritizedRequest wraps PoolIORequest and adds scheduling
  related fields.

More over, for NFS the mover itself is just an adapter for the
NFS code.

There is far too much wrapping going on. The encapsulation represented
by the above classes is no longer valid as internal state was exposed
in the refactoring that was done in dCache 1.9.11. The current situation
is difficult to understand and tightly coupled.

This patch eliminates PoolIORequest and the IoProcessable interface.

Target: trunk
Require-book: no
Require-notes: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5431/