Change webadmin alarms DAO to do lazy initialization for XML database
module: webadmin (alarms)

The current initialization of the web app uses a greedy startup of the underlying DAO for alarms;
if this is an XML plugin (the default), the file may not be there if the alarmserver has not already been started;
moreover, the pages are created regardless of whether the deployment is actually making use of the alarm server.

The code currently tries to work around this by waiting 30 seconds, and then providing a dummy NOP implementation
of the DAO API if the file is still not there.  This is not an optimal solution.

This patch changes the DAO implementation to lazily initialize the DataNucleus layer.  This also remedies the
situation (rare, but possible) in which the app is started with an XML file but in which the alarmserver not
only gets turned off but the file gets deleted.

Patch: http://rb.dcache.org/r/5407
Acked-by: Gerd
Target: trunk
Require-book: no
Require-notes: yes

RELEASE NOTES:
The case where an alarmserver is started against an already running dcache installation, or where it is stopped and
the xml file store is moved or deleted, is not correctly handled.