Add p2p attribute to MoverInfoMessage
module: core class pool, mover info

Part 1.

In the current billing logs and data tables there is no coherent way to track migration/pool-to-pool operations as distinct from door-initiated uploads and downloads.   This has been done partially at Fermilab (with respect to the text logs) by using a regex filter on the accounting entries; however, the plots generated from the database have never reflected this distinction.  Moreover, a change from 1.9.12+ to 2 which entailed switching to Http from dcap for the p2p protocol also brought with it a change to how the "initiator" is recorded (formerly, simply as <undefined>) which invalidates the matching in the Fermi script but which makes it less certain as to whether any general pattern will now catch the p2p cases.

This patch initiates a few changes which enable the tracking p2p transfers as distinct from door-initiated uploads and downloads.

Several of the relevant mover and pool messages have been modified to include a 'p2p' field, which is also included in the template for creating the accounting log lines (see billing.properties); the information is passed on by the pool when the billing message is sent.

Note: I realize that there is a patch pending which deprecates PoolIORequest.  I have posted this patch anyway, with the understanding that the appropriate modifications will be made when those changes are committed.

Acked-by: Gerd
Target: trunk
Patch: http://rb.dcache.org/r/5413
Require-book: no
Require-notes: yes

RELEASE NOTES:

These should be combined with those of the subsequent patches which actually implement the code for the plots to display this new data.  On the basis of this patch alone, no changes will be visible, though the MoverInfoMessage format could be altered to include the new p2p attribute (see the comments in the billing.properties file).

Notes: the new field has not been added to the default billing string format for backward compatibility.  It needs to be added by hand if you wish the logs to reflect the difference between p2p and door-initiated transfers.  The initiator field now begins with either "pool:" or "door:", respectively.

Testing done in conjunction with dependent patches.