Upgrade to Jetty 7.6
The version of Jetty we use is no longer supported. This patch
upgrades to the latest 7.6 release.

SSL support has been drastically updated in Jetty and as a
result our code for async GSI handling no longer works. I tried
to port the code to Jetty 7.6, but the changes are so substantial
that the port amounts to a rewrite.

Since the async connector never worked without issues, I decided
to drop support for the async connector instead. We may consider
adding it back in the future after moving to Jetty 8 or 9 and
possibly making better use of the JGlobus 2 architecture.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5253/