pool: Fix CPU spin in Netty based transfer services
Netty 3.5.7 modified the channel factory to no longer lazily
instatiate the worker threads at channel registration. Instead
it starts the workers in its constructor.

In some cases - and I haven't figured out exactly what those
cases are - the resulting worker begins to spin on the empty
selector. This may be a problem limited to Mac OS X, or it
may be a more generic problem.

I tried to write a test program to reproduce the problem, but
the problem is not easy to reproduce. For this reason it is
also difficult to report this problem to the Netty project.

As a solution this patch adds lazy instantiation of the channel
factory. Althoug one may consider this a woraround, the change
is necessary to avoid instantiating the worker unless we need
them. The premis that creating the channel factory early is cheap
no longer holds.

Target: trunk
Request: 2.6
Require-book: no
Require-notes: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5466/