Add ANSI output for admin shell help text
Adds pleasant looking formatting using bold and underlined
text for help output of some admin shell commands. Support
for more commands may be added in the future.

One of the motivations for introducing annotated commands as
a replacement to the old ac_ methods was so we could make
programatic use of the additional information captured by
the annotations.

This patch adds one such use, namely the ability to output
help text formatted for an ANSI console. I have copied the
output format of man pages, using bold for headings and
literal strings, and underlined text for value placeholders.

The short summary is that the help output of some commands
in the admin is now much prettier.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5469/