pool: Make access time update non-blocking
Addresses a performance issue when reading files from a pool.

Patch http://rb.dcache.org/r/4568/ modified the pool to let access
time be updated whenever a file is read. The update is however blocking,
which means that the notification thread in the pool is blocked until
a reply from PnfsManager is received. Given that we do not use the
reply for anything, we can just as well turn the access time update
into a non-blocking notification.

Target: trunk
Request: 2.6
Require-notes: no
Require-book: no
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5472/