SRM: don't talk to SpaceManager when disabled
There is a configuration value describing whether or not space-manager
is enabled.  Currently, SRM sends messages to space-manager even when
space-manager is disabled.  Because of this, sites are required to run
space-manager, even when it is disabled.

This patch alters the SRM so that if space-manager is disabled then not
communication with space-manager is attempted.  Under this circumstance,
SRM will report any space-token information for srmLs operations and
space-specific operations will fail with a friendly error message.

Conflicts:
	modules/dcache/src/main/java/diskCacheV111/srm/dcache/Storage.java
	modules/dcache/src/main/resources/diskCacheV111/srm/srm.xml
	skel/share/services/srm.batch