SRM: log client disconnecting at debug level
It is possible for the client to "give up" on a connection during the
SSL/GSI handshake exchange.  For example, this can happen if the client
doesn't trust the server's certificate.

When this happens, an EOFException is logged; for example:

30 Apr 2013 19:31:59 (SRM-zitpcx6184) [] java.io.EOFException

This message is both unintelligible and unnecessary, as dCache is
behaving correctly under the circumstances.

This patch updates the message to be meaningful, including the client's
IP address and port number in the diagnostic context. It also reduces
the log-level to debug.

Target: master
Request: 2.6
Committed: master@5e2edae4cdd6a10d2bdb499071edac4fc06bc74e
Patch: http://rb.dcache.org/r/5534/
Acked-by: Gerd Bahrmann