gsidcap: fix regression introduced by 969903b5
The patch entitled "a step towards configurable cell names for gPlazma
and billing", committed master@969903b5 added support in various doors
for configuring the gPlazma and billing cell names.

Unfortunately, the part touching the gsidcap door was broken as the
resulting code attempts to create the LoginStrategy before the gPlazma
CellStub is defined.  This results in an NPE any time a client attempts
to connect to dCache.

This patch fixes that issue by creating the LoginStrategy after the
required gPlazma CellStub is defined.

Patch: http://rb.dcache.org/r/5537/
Acked-by: Gerd Behrmann
Target: master