dcache: fix DoorRequestInfoMessage so that owner is DN if there is public credential
With gPlazma 2, the XACML plugin no longer has an ad hoc data structure (like KPWD) from which the DN of the user can be derived on the door message, because the authorized principals include only UserName, UID and GID principals.  However, our CMS Tier1 users say they require this database field to reflect the DN, and not the role name which ends up appearing there.

This fix is simple: the best effort to find a DN and make the "owner" field carry it now includes a search of the certificate chain.

Target: 2.2
Patch: http://rb.dcache.org/r/5536
Require-book: no
Require-notes: yes
Acked-by: Dmitry