webdav: Bootstrap for directory listing
Switches the entire styling of the webdav door to Bootstrap.

I have chosen a style that is very toned down, with minimal use of colors.
This is more suited for a default style, IMO.

The default template file is much cleaner now, which should simplify
customization. As always, sites can customize the style by copying
html.stg, modify it and set webdav.templates.html to point to the
copy.

This patch also fixes a rendering bug in the breadcrumbs.

jQuery and the tablesorter plugins are updated to the latest versions.
Bootstrap is used to theme tablesorter.

StringTemplate is updated to version 4.0.8.

To avoid that an Internet connection is required, I have chosen to
not use a CDN for the third party packages. I have embedded jquery,
bootstrap and tablesorter in the patch. I have included more or less
the entire package, even though only a subset is used in our
template.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7045/